requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
crispy-bootstrap4==2022.1
cryptography==41.0.3
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.5
django-allauth==0.55.2
django-countries==7.5.1
django-crispy-forms==2.0
django-filter==23.2
django-simple-history==3.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
idna==3.4
oauthlib==3.2.2
Pillow==10.0.0
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sentry-sdk==1.39.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.6.0
